requirements.txt
ollama
requests
python-dotenv